requirements.txt
asn1crypto==0.24.0
astroid==1.6.6
attrs==17.4.0
Automat==0.6.0
awscli==1.16.169
beautifulsoup4==4.7.1
blinker==1.4
botocore==1.12.159
certifi==2019.3.9
chardet==3.0.4
click==6.7
cloud-init==19.1
colorama==0.3.9
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
distro-info===0.18ubuntu0.18.04.1
Django==2.0.2
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
ikp3db==1.1.4
incremental==16.10.1
isort==4.3.20
jedi==0.11.1
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.4
jsonpatch==1.16
jsonpointer==1.10
jsonrpclib==0.1.7
jsonrpclib-pelix==0.4.0
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.1
lxml==4.3.4
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.4.3
oauthlib==2.0.6
olefile==0.45.1
PAM==0.4.2
parso==0.1.1
pbr==5.2.1
pika==1.0.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.5
pymongo==3.8.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.1
pyxdg==0.25
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
soupsieve==1.9.1
ssh-import-id==5.7
stevedore==1.30.1
systemd-python==234
tinysegmenter==0.3
tldextract==2.2.1
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.1
zope.interface==4.3.2
